% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{col_means}
\alias{col_means}
\title{Column_means}
\usage{
col_means(data_frame)
}
\arguments{
\item{data_frame}{}
}
\value{
A numeric vector containing the column means.
}
\description{
This function calculates the mean of each column in a data frame.
}
\examples{
sample_df <- data.frame(A = c(1, 2, 3), B = c(4, 5, 6), C = c(7, 8, 9))
col_means(sample_df)
}
