% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{count_na}
\alias{count_na}
\title{Count NA}
\usage{
count_na(vec)
}
\arguments{
\item{vec}{}
}
\value{
An integer indicating the number of NA values.
}
\description{
This function counts the number of NA values in a vector using a for-loop.
}
\examples{
sample_vector <- c(1, 2, NA, 4, NA, 6, NA)
count_na(sample_vector)
}
