% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{col_means}
\alias{col_means}
\title{Column Means Function}
\usage{
col_means(data_frame)
}
\arguments{
\item{data_frame}{a data frame}
}
\value{
a vector of numeric column means
}
\description{
Column Means Function
}
\details{
It allows you to compute the means of columns in a given data frame
}
\examples{
ex_data1 <- data.frame(a = c(1, 2, 3, 4), b = c(5, 6, 7, 8))
col_means(ex_data1)

ex_data2 <- data.frame(a = c(2, 4, 6, NA), b = c(8, 10, 12, 14))
col_means(ex_data2)
}
