% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{col_means}
\alias{col_means}
\title{Column Means}
\usage{
col_means(df)
}
\arguments{
\item{df}{is a dataframe}
}
\value{
a named vector of the means of each column of df
}
\description{
Column Means
}
\examples{
col_means(mtcars)
}
