% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{col_means}
\alias{col_means}
\title{col_means()}
\usage{
col_means(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
A vector of means from a data frame
}
\description{
This function reads a data frame and returns a vector of the averages for each column by using a for loop
}
\author{
Ronan Tanona
}
