% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{count_na}
\alias{count_na}
\title{NA Count Function}
\usage{
count_na(vector)
}
\arguments{
\item{vector}{a vector}
}
\value{
an integer count of NA values
}
\description{
NA Count Function
}
\details{
It allows you to count the number of NA values that appears in a given vector
}
\examples{
ex_vector1 <- c(1, 2, NA, 4, NA, 6)
count_na(ex_vector1)
}
