% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{count_na}
\alias{count_na}
\title{Count NAs}
\usage{
count_na(vec)
}
\arguments{
\item{vec}{is a vector}
}
\value{
the number of NAs in vec
}
\description{
Count NAs
}
\examples{
count_na(c(1,1,NA,2,NA))
}
